Refactor delegator to be a subtrait of NameInterpreter called DelegatingNameInterpreter.
