+++ namerd/main/src/main/scala/io/buoyant/namerd/NamerdAdmin.scala
@@ -4,7 +4,7 @@ import com.twitter.finagle.Service
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.server.handler.ResourceHandler
 import io.buoyant.admin.{Admin, App, ConfigHandler, StaticFilter}
-import io.buoyant.linkerd.admin.names.DelegateApiHandler
+import io.buoyant.admin.names.DelegateApiHandler
 
 class NamerdAdmin(app: App, config: NamerdConfig, namerd: Namerd) extends Admin(app) {
   override def allRoutes: Seq[(String, Service[Request, Response])] = {
@@ -19,7 +19,7 @@ class NamerdAdmin(app: App, config: NamerdConfig, namerd: Namerd) extends Admin(
       localFilePath = "admin/src/main/resources/io/buoyant/admin"
     )),
     "/" -> new DtabListHandler(namerd.dtabStore),
-    "/delegator.json" -> new DelegateApiHandler(namerd.namers.toSeq),
+    "/delegator.json" -> new DelegateApiHandler(ns => ConfiguredDtabNamer(namerd.extractDtab(ns), namerd.namers.toSeq)),
     "/dtab/" -> new DtabHandler(namerd.dtabStore)
   )
 }
